+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+
+# 0.4.0
+
+- Add packages
+- Refactor the standard library as multiple packages
+
+# 0.3.3
+
+- Fix the operator precedence of ~
+
 # 0.3.2
 
 - Fix confusing printing of lambdas